ai-agent-api/pyproject.toml
[tool.poetry]
name = "ai-agent-api"
version = "1.0.0"
description = "Enterprise-grade API service for Claude Code Agent SDK with session management and business logic"
authors = ["KubeMind Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
# Core Framework
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"

# Redis & Caching
redis = {extras = ["hiredis"], version = "^5.0.1"}
hiredis = "^2.3.2"

# Celery
celery = "^5.3.6"

# Claude SDK
claude-agent-sdk = "^0.1.4"

# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

# Monitoring
prometheus-client = "^0.19.0"
structlog = "^24.1.0"
sentry-sdk = "^1.40.0"

# Utilities
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
croniter = "^2.0.1"
weasyprint = "^60.2"
httpx = "^0.26.0"

# Kubernetes (for MCP tools)
kubernetes = "^29.0.0"
pyjwt = "^2.10.1"
email-validator = "^2.3.0"
psutil = "^7.1.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
faker = "^22.6.0"

# Development
black = "^24.1.1"
ruff = "^0.1.15"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
aiosqlite = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
